Move translation class to its own file
